update todo
note: seeds are not practical:
if we were to convert the binary for a 60x100 map into base64, it would be 6000 bin characters, so 1000 base64 characters, so the url would have an additional 1000characters. there are no seed options in the Math.random library in js, and other libraries exist but are a pain. we would need a server for a dynamic website if we want to pair user generated maps with shortened links, and no backend is possible on a static github pages website. so tl;dr: no seeds :(